Phase 1 polish: move agent.rs tests to agent_tests.rs + tighten comments
Two cleanups before this PR ships:

1. Move the inline `#[cfg(test)] mod tests` block from
   `app/src/persistence/agent.rs` into `app/src/persistence/agent_tests.rs`
   and replace it with the `#[cfg(test)] #[path = "agent_tests.rs"]
   mod tests;` shim per the WARP.md convention. All 8
   `select_conversations_to_evict` tests still run; nextest confirms
   `persistence::agent::tests` finds them under their original module
   path.

2. Comment pass across the files touched by this PR's substantive
   changes. The PR iteration accumulated lecture-y multi-paragraph
   rationales; this commit collapses them to 1-3 lines each while
   preserving "why" notes, documented invariants, and doc comments on
   `pub`/`pub(super)` items. No behaviour changes, no renames, no
   signature changes; comments only.

   Files trimmed:
   - `app/src/pane_group/mod.rs` (`hydrate_task_backed_hidden_child_pane`,
     `attempt_remote_child_hydration`,
     `decide_remote_child_hydration_action`,
     `attach_ambient_session_and_maybe_tombstone`,
     `apply_existing_ambient_task_to_pane`,
     `hydrate_remote_child_transcript_in_place`,
     `RemoteChildHydrationAction`, `PendingRemoteChildHydration`,
     `pending_remote_child_hydrations`,
     `pending_ambient_restoration_subscription_installed`).
   - `app/src/ai/blocklist/history_model.rs`
     (`MAX_HISTORICAL_CONVERSATIONS`,
     `merge_cloud_tasks_into_existing_conversation`).
   - `app/src/persistence/agent.rs` (`MAX_PERSISTED_CONVERSATION_COUNT`,
     `select_conversations_to_evict`).
   - `crates/persistence/src/model.rs` (`task_is_root_shaped`,
     `optimistic_stub_task_id`, `tasks_for_restore`,
     `into_tasks_for_restore`, `is_restorable`).
   - `app/src/ai/blocklist/history_model/conversation_loader.rs`
     (`convert_persisted_conversation_to_ai_conversation_with_metadata`,
     eager-hydration block in `initialize_historical_conversations`).

Validation: cargo fmt -p warp -p persistence + cargo clippy -p warp
--all-targets --features local_fs -- -D warnings + cargo clippy -p
persistence --tests --all-features -- -D warnings + cargo nextest run
-p warp persistence::agent::tests pane_group:: ai::blocklist::history_model::tests
(143 passed) + cargo nextest run -p persistence (16 passed). Test
counts unchanged.
